Disabled all of PhysX and several parts of the code that interfered with testing Lumen and Nanite.

+++ Unreal/CarlaUnreal/Plugins/Carla/Source/Carla/Recorder/CarlaRecorderHelpers.cpp
@@ -66,9 +66,9 @@ void WriteFString(std::ostream &OutFile, const FString &InObj)
 // read binary data to FVector
 void ReadFVector(std::istream &InFile, FVector &OutObj)
 {
-  ReadValue<float>(InFile, OutObj.X);
-  ReadValue<float>(InFile, OutObj.Y);
-  ReadValue<float>(InFile, OutObj.Z);
+  ReadValue<double>(InFile, OutObj.X);
+  ReadValue<double>(InFile, OutObj.Y);
+  ReadValue<double>(InFile, OutObj.Z);
 }
 
 // read binary data to FTransform
